ft_strlcpy.c

+++ ft_itoa.c
@@ -1,6 +1,18 @@
-#include <libft.h>
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_itoa.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: rrivelin <rrivelin@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/09/13 20:04:52 by rrivelin          #+#    #+#             */
+/*   Updated: 2022/09/13 20:15:07 by rrivelin         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-static char *convert_itoa(char *str,
+#include "libft.h"
+
+static char	*convert_itoa(char *str,
 								size_t size,
 								unsigned int num,
 								unsigned int is_negative)
@@ -11,38 +23,36 @@ static char *convert_itoa(char *str,
 		str[size] = (num % 10) + 48;
 		num /= 10;
 	}
-	if(is_negative)
+	if (is_negative)
 	{
 		str[0] = '-';
 	}
-	return(str);
-	
+	return (str);
 }
 
-size_t ft_numlen(int num)
+size_t	ft_numlen(int num)
 {
-    size_t len;
+	size_t	len;
 
-    len = 1;
-
-    if (num < 0)
-    {
-        len++;
+	len = 1;
+	if (num < 0)
+	{
+		len++;
 		num /= 10;
-    }
-    while (num)
-    {
-        num /= 10;
-        len++;
-    }
+	}
+	while (num)
+	{
+		num /= 10;
+		len++;
+	}
 	return (len);
 }
 
-char *ft_itoa(int n)
+char	*ft_itoa(int n)
 {
-	size_t n_digits;
-	unsigned int is_negative;
-	char	*str;/* condition */
+	size_t	n_digits;
+	long	is_negative;
+	char	*str;
 
 	n_digits = ft_numlen(n);
 	is_negative = 0;
@@ -54,8 +64,7 @@ char *ft_itoa(int n)
 	str = malloc(sizeof(char) * (n_digits + 1));
 	if (str == NULL)
 	{
-		return(NULL);
+		return (NULL);
 	}
 	return (convert_itoa(str, n_digits, (unsigned int)n, is_negative));
-
 }
